+++ attention/prompt_flash_attention/op_host/prompt_flash_attention_tiling.h
@@ -36,18 +36,18 @@ namespace optiling {
 
 struct PFAShapeInfo {
     uint32_t b = 0;
-    uint32_t n = 0;
-    uint32_t s = 0;
+    uint64_t n = 0;
+    uint64_t s = 0;
     uint32_t d = 0;
-    uint32_t h = 0;
+    uint64_t h = 0;
     uint32_t t = 0;
 };
 
 BEGIN_TILING_DATA_DEF(PromptAttentionBaseParams)
     TILING_DATA_FIELD_DEF(uint32_t, batchSize);
-    TILING_DATA_FIELD_DEF(uint32_t, headNumSize);
+    TILING_DATA_FIELD_DEF(uint64_t, headNumSize);
     TILING_DATA_FIELD_DEF(uint32_t, seqSize);
-    TILING_DATA_FIELD_DEF(uint32_t, headSize);
+    TILING_DATA_FIELD_DEF(uint64_t, headSize);
     TILING_DATA_FIELD_DEF(float, scaleValue);
     TILING_DATA_FIELD_DEF(int32_t, preTokens);
     TILING_DATA_FIELD_DEF(int32_t, nextTokens);
@@ -60,7 +60,7 @@ BEGIN_TILING_DATA_DEF(PromptAttentionBaseParams)
     TILING_DATA_FIELD_DEF(uint32_t, prefixSeqInnerSize);
     TILING_DATA_FIELD_DEF(uint32_t, usePseShift);
     TILING_DATA_FIELD_DEF(uint32_t, useMask);
-    TILING_DATA_FIELD_DEF(uint32_t, headNumRatio);
+    TILING_DATA_FIELD_DEF(uint64_t, headNumRatio);
     TILING_DATA_FIELD_DEF(uint32_t, attenMaskElemType);
     TILING_DATA_FIELD_DEF(uint32_t, pseShiftTypeByteNum);
     TILING_DATA_FIELD_DEF(uint32_t, pseMaskMaxSize);
@@ -111,12 +111,12 @@ REGISTER_TILING_DATA_CLASS(PromptAttentionBaseParamsOp, PromptAttentionBaseParam
 
 BEGIN_TILING_DATA_DEF(PromptAttentionBaseApiBaseParams)
     TILING_DATA_FIELD_DEF(uint32_t, batchSize);
-    TILING_DATA_FIELD_DEF(uint32_t, headNumSize);
-    TILING_DATA_FIELD_DEF(uint32_t, headSize);
+    TILING_DATA_FIELD_DEF(uint64_t, headNumSize);
+    TILING_DATA_FIELD_DEF(uint64_t, headSize);
     TILING_DATA_FIELD_DEF(uint32_t, maskTypeByteNum);
     
     TILING_DATA_FIELD_DEF(uint32_t, inputLayoutType);
-    TILING_DATA_FIELD_DEF(uint32_t, kvHeadNumSize);
+    TILING_DATA_FIELD_DEF(uint64_t, kvHeadNumSize);
     TILING_DATA_FIELD_DEF(uint32_t, maxSeqLen);
     TILING_DATA_FIELD_DEF(uint32_t, maxKvSeqLen);
     TILING_DATA_FIELD_DEF(uint32_t, totalQBlkNum);
@@ -591,16 +591,16 @@ protected:
                         const uint32_t typeByteSize);
     uint32_t CalculateL1SizeUsed(PromptFlashAttentionTilingData& tilingData, const uint32_t typeByteSize);
     bool CheckInputDimAndHeadNum(ContextParamsForPFATiling& contextKeyParams, uint32_t nQAttr, uint32_t nKVAttr);
-    bool SetTilingHeadNumRatio(ContextParamsForPFATiling& contextKeyParams, const int32_t* numQueryHeads,
-                               const int32_t* numKeyValueHeads, PromptFlashAttentionTilingData& tilingData);
+    bool SetTilingHeadNumRatio(ContextParamsForPFATiling& contextKeyParams, const int64_t* numQueryHeads,
+                               const int64_t* numKeyValueHeads, PromptFlashAttentionTilingData& tilingData);
     void PromptFlashAttentionInitOutputSplit(uint64_t totalSize, PromptFlashAttentionTilingData &tilingData,
                                              uint32_t curCoreNum);
     void PromptFlashAttentionInitSoftmaxLseOutputSplit(uint64_t totalSize, PromptFlashAttentionTilingData &tilingData);
     void Align(uint32_t &num);
     ge::graphStatus GetBasicShape(uint32_t &b, uint32_t &s, uint32_t &h, uint32_t &seqInnerSize,
                                 const gert::StorageShape *queryShape, const gert::StorageShape *keyShape, const uint32_t n);
-    ge::graphStatus GetBasicShape310P(uint32_t &b, uint32_t &bKV, uint32_t &s, uint32_t &h, uint32_t &seqInnerSize,
-                                      const gert::StorageShape *queryShape, const gert::StorageShape *keyShape, const uint32_t n,
+    ge::graphStatus GetBasicShape310P(uint32_t &b, uint32_t &bKV, uint32_t &s, uint64_t &h, uint32_t &seqInnerSize,
+                                      const gert::StorageShape *queryShape, const gert::StorageShape *keyShape, const uint64_t n,
                                       size_t actualLenDims, size_t actualLenDimsKV);
     ge::graphStatus GetBasicShape910B(uint32_t &b, uint32_t &s, uint32_t &h, uint32_t &seqInnerSize,
                                       const gert::StorageShape *queryShape, const gert::StorageShape *keyShape, const uint32_t n);
@@ -610,35 +610,35 @@ protected:
     bool CheckActualSeqLength(ContextParamsForPFATiling& contextKeyParams, uint32_t b, uint32_t sQ, uint32_t sKV,
                               const gert::Tensor* actualSeqLenQ, const gert::Tensor* actualSeqLenKV, InputLayout inLayout, PromptFlashAttentionTilingData& tilingData);
     bool CheckPseShiftTypeAndShape(ContextParamsForPFATiling& contextKeyParams, const gert::StorageShape *pseShiftShape,
-                                   uint32_t b, uint32_t n, uint32_t s1, uint32_t s2);
+                                   uint32_t b, uint64_t n, uint32_t s1, uint32_t s2);
     ge::graphStatus processPageAttentionInputFlag(ContextParamsForPFATiling& contextKeyParams);
     bool checkPAKeyValueDimsWhenBBH(ContextParamsForPFATiling& contextKeyParams, int32_t keyDim1, int32_t keyDim2, int32_t keyDim3, int64_t blockNumValid,
-        const int32_t* curBlockSize, int32_t h, int32_t headNumRatio);
+        const int32_t* curBlockSize, int64_t h, int64_t headNumRatio);
     bool checkPAKeyValueDimsWhenBNBD(ContextParamsForPFATiling& contextKeyParams, const gert::StorageShape* keyShape, const gert::StorageShape* valueShape, int32_t keyDim1, 
-                                int32_t keyDim2, int32_t keyDim3, int64_t blockNumValid, const int32_t* curBlockSize, int32_t h, int32_t n, int32_t headNumRatio);
+                                int32_t keyDim2, int32_t keyDim3, int64_t blockNumValid, const int32_t* curBlockSize, int64_t h, int64_t n, int64_t headNumRatio);
     bool checkPAKeyValueDimsWhenNZ(ContextParamsForPFATiling& contextKeyParams, const gert::StorageShape* keyShape, const gert::StorageShape* valueShape, int32_t keyDim1, 
-                                int32_t keyDim2, int32_t keyDim3, int64_t blockNumValid, const int32_t* curBlockSize, int32_t h, int32_t n, int32_t headNumRatio);
+                                int32_t keyDim2, int32_t keyDim3, int64_t blockNumValid, const int32_t* curBlockSize, int64_t h, int64_t n, int64_t headNumRatio);
     bool checkPABlockSizeAndBlockTable(ContextParamsForPFATiling& contextKeyParams, const gert::Tensor* actualSeqLenKV, const int32_t* curBlockSize, int64_t b);
     bool culActSeqLenParamsWhenPA(ContextParamsForPFATiling& contextKeyParams, const gert::Tensor* actualSeqLenKV, int64_t& blockNumValid, 
                                 int32_t& maxBlockNumPerBatch, int32_t tempBlockSize, const int32_t* curBlockSize, int64_t b);
     bool CheckPAKeyValueParams(ContextParamsForPFATiling& contextKeyParams, const gert::StorageShape* keyShape, const gert::StorageShape* valueShape, 
-                                int64_t blockNumValid, const int32_t* curBlockSize, int32_t n, int32_t h, int32_t headNumRatio);
+                                int64_t blockNumValid, const int32_t* curBlockSize, int64_t n, int64_t h, int64_t headNumRatio);
     bool CheckPASparseMode(ContextParamsForPFATiling& contextKeyParams);
     bool CheckPAWhenBaseApi(ContextParamsForPFATiling& contextKeyParams, const gert::Tensor* actualSeqLenQ, const gert::Tensor* actualSeqLenKV,
-                                   int32_t n, int32_t h, int32_t headNumRatio);
+                                   int64_t n, int64_t h, int64_t headNumRatio);
     bool CheckPATypeAndShape(ContextParamsForPFATiling& contextKeyParams, const gert::Tensor* actualSeqLenKV,
-                                   int32_t b, int32_t n, int32_t h, int32_t headNumRatio);
+                                   int32_t b, int64_t n, int64_t h, int64_t headNumRatio);
     bool CheckAttenMaskShape(ContextParamsForPFATiling& contextKeyParams, const int32_t* sparseMode, const gert::StorageShape* attenMaskShape,
-                             uint32_t sQ, uint32_t sK, uint32_t batchSize);
+                             uint64_t sQ, uint64_t sK, uint32_t batchSize);
     bool CheckPAAntiquantSupportScenarios(ContextParamsForPFATiling& contextKeyParams, PromptFlashAttentionTilingData& tilingData);
     bool CheckPerchannelAntiquantParamsShape(ContextParamsForPFATiling& contextKeyParams, const gert::StorageShape* antiquantScaleShape, const gert::StorageShape* antiquantOffsetShape, 
-                                             const uint32_t n, const uint32_t d, const uint32_t h, uint32_t paramFirstDim) const;
+                                             const uint64_t n, const uint32_t d, const uint64_t h, uint32_t paramFirstDim) const;
     bool CheckPerchannelBSNDParamsShape(ContextParamsForPFATiling& contextKeyParams, const gert::StorageShape* antiquantScaleShape, const gert::StorageShape* antiquantOffsetShape, 
-                                        const uint32_t n, const uint32_t d, uint32_t paramFirstDim) const;
+                                        const uint64_t n, const uint32_t d, uint32_t paramFirstDim) const;
     bool CheckAntiquantParamsShape(ContextParamsForPFATiling& contextKeyParams, const gert::StorageShape* antiquantScaleShape,
-                                   const gert::StorageShape* antiquantOffsetShape, const uint32_t n, const uint32_t d, const uint32_t h,
+                                   const gert::StorageShape* antiquantOffsetShape, const uint64_t n, const uint32_t d, const uint64_t h,
                                    PromptFlashAttentionTilingData& tilingData);
-    ge::graphStatus CheckPostQuantParams(const ContextParamsForPFATiling& contextKeyParams, uint32_t h, uint32_t n) const;
+    ge::graphStatus CheckPostQuantParams(const ContextParamsForPFATiling& contextKeyParams, uint64_t h, uint64_t n) const;
     ge::graphStatus PromptFlashAttentionCVDiffSetTensorSize(PromptFlashAttentionTilingData& tilingData,
         PromptAttentionSingleCoreTensorSize& tensorSize, uint32_t sOuterFactor,
         uint32_t sInnerFactor, uint32_t softmaxSOuterFactor);
@@ -653,7 +653,7 @@ protected:
         int64_t ubSize, uint32_t typeByteSize, uint32_t maskTypeSize);
     void InferTilingMod(const ContextParamsForPFATiling& contextKeyParams, const std::vector<int64_t>& actualSeqLengths, const std::vector<int64_t>& actualSeqLengthsKV,
         uint32_t actualSeqArrayLen, uint32_t hDivN, uint32_t seqInnerSize, int32_t sparseModeVal);
-    ge::graphStatus AdjustCVTiling(uint32_t hDivN, uint32_t n, int64_t middle_actualSeqLengths,
+    ge::graphStatus AdjustCVTiling(uint64_t hDivN, uint64_t n, int64_t middle_actualSeqLengths,
         int64_t ubSize, int64_t l1Size, int64_t l0CSize, uint32_t maskElemSize,
         uint32_t& sOuterFactor, uint32_t& sInnerFactor, PromptFlashAttentionTilingData& tilingData);
     ge::graphStatus AdjustCVTilingCVDiff(int64_t ubSize, int64_t l1Size, int64_t l0CSize,
